api: Split up api/flash into smaller files
