+++ src/api/flash/utils.js
@@ -1,0 +1,24 @@
+/* chrysalis-flash -- Keyboard flash helpers for Chrysalis
+ * Copyright (C) 2018-2022  Keyboardio, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { logger } from "@api/log";
+
+export const delay = (ms) => new Promise((res) => setTimeout(res, ms));
+
+export const toStep = (callback) => async (step) => {
+  logger("flash").info("executing step", { step: step });
+  return callback(step);
+};
