+++ src/api/hardware-softhruf-splitography/index.js
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { DFUProgrammer } from "@api/flash";
+import { DFUProgrammer } from "@api/flash/DFUProgrammerFlasher";
 import Keymap from "./components/Keymap";
 
 const Splitography = {
