+++ src/renderer/screens/Preferences/keyboard/AdvancedKeyboardPreferences.js
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { FocusCommands } from "@api/flash";
+import { FocusCommands } from "@api/flash/FocusCommands";
 import Backdrop from "@mui/material/Backdrop";
 import Button from "@mui/material/Button";
 import CircularProgress from "@mui/material/CircularProgress";
